the pizza bounce
getting down with a whole pizza is my lifestyle, also how i learned a
bit more animation with CSS today.

this seems pretty simple (it probably is), but it took me a while to
understand how it was all working together. because i’m so new at this,
it’s a lot of reading just to understand what’s doing what, and why.

but now i can make a bouncing pizza! a generally cool life skill i
guess.